fix+feat: 7 issues — prefix-sum masks, ValueAccessor, ARM portability, v220 compression, 3 TinyVDB bugs
- Mask{N,W}: add prefix-sum NTuple for O(1) count_on_before (was O(W))
- ValueAccessor{T}: mutable cache for leaf/I1/I2 nodes, 5-8x speedup for trilinear
- Binary.jl: memcpy-based _unaligned_load for ARM portability (was unsafe_load)
- TinyVDB read_grid_compression: propagate header.is_compressed for v220 files
- TinyVDB read_compressed_data: add abs(chunk_size) cross-validation
- TinyVDB read_metadata: use read_u32 for size prefixes (was read_i32)
- Coordinates.jl volume(BBox): return Int128 to avoid overflow
- File.jl: @warn for unsupported grid value types instead of silent skip

920 tests pass.

+++ test/test_masks.jl
@@ -176,7 +176,7 @@
         @test count_on_before(m, 127) == 3  # Bits 0, 63, 64 before 127
 
         # Verify consistency: count_on_before(m, i) + 1 == position in on_indices for bit i
-        words = (0b10110001,)  # bits 0, 4, 5, 7
+        words = (UInt64(0b10110001),)  # bits 0, 4, 5, 7
         m = Mask{64,1}(words)
         indices = collect(on_indices(m))
         for (pos, idx) in enumerate(indices)
